# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MEDI-IOT EyeCare API",
    "description": "\n## MEDI-IOT EyeCare — 안과 AI 진단 플랫폼\n\n### 주요 기능\n- **환자 관리**: 등록, 조회, PII 암호화 저장\n- **검사 기록**: OCT, 안저, 시야 검사 등 다양한 검사 유형 지원\n- **AI 진단**: shared-libraries Orchestrator (CONSENSUS 전략) 기반\n  - FAST 모델: gemma-4-e4b (초안 생성)\n  - HEAVY 모델: gemma-4-26b-a4b (의료 정확성 검토)\n  - OntologyValidator: MEDICAL 도메인 규칙 자동 검증\n\n### 설계 원칙\n- 환자 PII는 암호화하여 저장 (응답 시 마스킹)\n- 의료 안전을 위해 CONSENSUS 전략 사용 (두 모델이 합의해야 PASS)\n- Circuit Breaker: 최대 2회 반복 후 현재 최선 결과 반환\n    ",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/patients/": {
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "환자 등록",
        "description": "환자를 등록합니다.\n\n- `patient_code`는 병원 내부 번호로 **고유값**이어야 합니다\n- `name`은 AES-256으로 암호화하여 저장됩니다\n- 응답의 `name_masked`는 첫 글자만 노출 (예: 홍길동 → 홍**)\n- `primary_diagnosis_code`는 ICD-10 형식 검증 (예: `H36.0`)",
        "operationId": "create_patient_api_v1_patients__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "등록된 환자 정보 (이름은 마스킹 처리)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "환자 목록 조회",
        "description": "환자 목록을 페이징하여 조회합니다.",
        "operationId": "list_patients_api_v1_patients__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientResponse"
                  },
                  "title": "Response List Patients Api V1 Patients  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "환자 조회",
        "description": "UUID 또는 `patient_code`로 환자를 조회합니다.\n\n- `patient_id` 파라미터에 UUID 또는 `P123456` 형식 모두 허용\n- `exam_count`: 해당 환자의 전체 검사 기록 수",
        "operationId": "get_patient_api_v1_patients__patient_id__get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "환자 UUID 또는 patient_code (예: P123456)",
              "examples": [
                "P123456"
              ],
              "title": "Patient Id"
            },
            "description": "환자 UUID 또는 patient_code (예: P123456)"
          }
        ],
        "responses": {
          "200": {
            "description": "환자 상세 정보 + 검사 횟수",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "patients"
        ],
        "summary": "환자 비활성화 (soft delete)",
        "description": "환자를 비활성화합니다 (실제 삭제 아님 — 의료 데이터 보존 의무).",
        "operationId": "deactivate_patient_api_v1_patients__patient_id__delete",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/exams": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "환자별 검사 목록",
        "description": "특정 환자의 전체 검사 기록을 조회합니다.",
        "operationId": "get_patient_exams_api_v1_patients__patient_id__exams_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExamResponse"
                  },
                  "title": "Response Get Patient Exams Api V1 Patients  Patient Id  Exams Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExamResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "검사 UUID"
          },
          "patient_id": {
            "type": "string",
            "title": "Patient Id",
            "description": "환자 UUID"
          },
          "exam_type": {
            "type": "string",
            "title": "Exam Type",
            "description": "검사 종류"
          },
          "exam_date": {
            "type": "string",
            "format": "date",
            "title": "Exam Date",
            "description": "검사 날짜"
          },
          "icd_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icd Code",
            "description": "ICD-10 코드"
          },
          "iop_left": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iop Left",
            "description": "좌안 안압 (mmHg)"
          },
          "iop_right": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iop Right",
            "description": "우안 안압 (mmHg)"
          },
          "visual_acuity_left": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual Acuity Left",
            "description": "좌안 시력"
          },
          "visual_acuity_right": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual Acuity Right",
            "description": "우안 시력"
          },
          "raw_findings": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Findings",
            "description": "검사 소견 원문"
          },
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary",
            "description": "LLM 생성 요약 (최대 500자)"
          },
          "report_status": {
            "type": "string",
            "title": "Report Status",
            "description": "보고서 상태: pending | generating | completed | failed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "등록일시 (UTC)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "exam_type",
          "exam_date",
          "icd_code",
          "iop_left",
          "iop_right",
          "visual_acuity_left",
          "visual_acuity_right",
          "raw_findings",
          "ai_summary",
          "report_status",
          "created_at"
        ],
        "title": "ExamResponse",
        "example": {
          "ai_summary": "당뇨망막병증 비증식성 중등도 소견. 추가 OCT 권장.",
          "created_at": "2026-05-09T10:35:00Z",
          "exam_date": "2026-05-09",
          "exam_type": "fundus",
          "icd_code": "H36.0",
          "id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
          "iop_left": 14.5,
          "iop_right": 15.2,
          "patient_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
          "raw_findings": "우안: 황반 주위 점상출혈 및 경성삼출물 다수 관찰.",
          "report_status": "completed",
          "visual_acuity_left": "0.8",
          "visual_acuity_right": "0.7"
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientCreate": {
        "properties": {
          "patient_code": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "Patient Code",
            "description": "병원 내부 환자 번호 (3~20자, 고유값)",
            "examples": [
              "P123456"
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "환자 이름 (AES-256 암호화 저장, 응답 시 마스킹)"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth",
            "description": "생년월일 (YYYY-MM-DD)",
            "examples": [
              "1975-03-15"
            ]
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(male|female|other)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender",
            "description": "성별: male | female | other"
          },
          "primary_diagnosis_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Diagnosis Code",
            "description": "주 진단 ICD-10 코드 (예: H36.0 당뇨망막병증)",
            "examples": [
              "H36.0"
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "임상 메모 (암호화 저장 대상 아님)"
          }
        },
        "type": "object",
        "required": [
          "patient_code"
        ],
        "title": "PatientCreate",
        "example": {
          "date_of_birth": "1975-03-15",
          "gender": "male",
          "name": "홍길동",
          "notes": "당뇨병 10년 이력, 정기 안저 검사 중",
          "patient_code": "P123456",
          "primary_diagnosis_code": "H36.0"
        }
      },
      "PatientResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "환자 UUID"
          },
          "patient_code": {
            "type": "string",
            "title": "Patient Code",
            "description": "병원 내부 환자 번호"
          },
          "name_masked": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Masked",
            "description": "마스킹된 이름 (첫 글자 + *)"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth",
            "description": "생년월일"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender",
            "description": "성별"
          },
          "primary_diagnosis_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Diagnosis Code",
            "description": "주 진단 ICD-10 코드"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "활성 여부 (false = soft delete)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "등록일시 (UTC)"
          },
          "exam_count": {
            "type": "integer",
            "title": "Exam Count",
            "description": "검사 기록 수",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_code",
          "is_active",
          "created_at"
        ],
        "title": "PatientResponse",
        "example": {
          "created_at": "2026-05-09T10:30:00Z",
          "date_of_birth": "1975-03-15",
          "exam_count": 3,
          "gender": "male",
          "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
          "is_active": true,
          "name_masked": "홍**",
          "patient_code": "P123456",
          "primary_diagnosis_code": "H36.0"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}